+++ test/readFromGraph.ts
