+++ test/tests.ts
@@ -53,6 +53,6 @@ import './isEqual';
 import './cloneDeep';
 import './assign';
 import './environment';
-import './graph_write';
-import './graph_read';
+import './writeToGraph';
+import './readFromGraph';
 import './GraphStore';
