+++ test/writeToGraph.ts
